Move grid square and foil hole endpoints
